--- packages/protocol/engine/clock.go
@@ -1,62 +1,0 @@
-package engine
-
-import (
-	"time"
-
-	"github.com/iotaledger/goshimmer/packages/core/module"
-	"github.com/iotaledger/hive.go/runtime/event"
-)
-
-// region Clock ////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// Clock is a component that provides different notions of time for the Engine.
-type Clock interface {
-	// Acceptance returns a notion of time that is anchored to the latest accepted block.
-	Acceptance() RelativeClock
-
-	// Confirmation returns a notion of time that is anchored to the latest confirmed block.
-	Confirmation() RelativeClock
-
-	// Interface embeds the required methods of the module.Interface.
-	module.Interface
-}
-
-// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// region RelativeClock ////////////////////////////////////////////////////////////////////////////////////////////////
-
-// RelativeClock is a time value that monotonically advances with the system clock but that is anchored to a specific
-// point in time.
-type RelativeClock interface {
-	// Time returns the time that this clock is anchored to.
-	Time() time.Time
-
-	// RelativeTime returns.
-	RelativeTime() time.Time
-}
-
-// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// region ClockEvents //////////////////////////////////////////////////////////////////////////////////////////////////
-
-// ClockEvents is a collection of events that are triggered by the Clock.
-type ClockEvents struct {
-	// AcceptanceTimeAnchorUpdated is triggered when the Acceptance Tangle Time advances.
-	AcceptanceTimeAnchorUpdated *event.Event1[time.Time]
-
-	// ConfirmationTimeAnchorUpdated is triggered when the Confirmation Tangle Time advances.
-	ConfirmationTimeAnchorUpdated *event.Event1[time.Time]
-
-	// LinkableCollection is a generic trait that allows to link multiple collections of events together.
-	event.Group[ClockEvents, *ClockEvents]
-}
-
-// NewClockEvents contains the constructor of the ClockEvents object (it is generated by a generic factory).
-var NewClockEvents = event.CreateGroupConstructor(func() (newEvents *ClockEvents) {
-	return &ClockEvents{
-		AcceptanceTimeAnchorUpdated:   event.New1[time.Time](),
-		ConfirmationTimeAnchorUpdated: event.New1[time.Time](),
-	}
-})
-
-// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
