+++ packages/protocol/engine/clock/events.go
@@ -1,0 +1,26 @@
+package clock
+
+import (
+	"time"
+
+	"github.com/iotaledger/hive.go/runtime/event"
+)
+
+type Events struct {
+	// AcceptanceTimeAnchorUpdated is triggered when the Acceptance Tangle Time advances.
+	AcceptanceTimeAnchorUpdated *event.Event1[time.Time]
+
+	// ConfirmationTimeAnchorUpdated is triggered when the Confirmation Tangle Time advances.
+	ConfirmationTimeAnchorUpdated *event.Event1[time.Time]
+
+	// LinkableCollection is a generic trait that allows to link multiple collections of events together.
+	event.Group[Events, *Events]
+}
+
+// NewEvents contains the constructor of the Events object (it is generated by a generic factory).
+var NewEvents = event.CreateGroupConstructor(func() (newEvents *Events) {
+	return &Events{
+		AcceptanceTimeAnchorUpdated:   event.New1[time.Time](),
+		ConfirmationTimeAnchorUpdated: event.New1[time.Time](),
+	}
+})
