+++ packages/protocol/engine/engine.go
@@ -9,6 +9,7 @@ import (
 	"github.com/pkg/errors"
 
 	"github.com/iotaledger/goshimmer/packages/core/module"
+	"github.com/iotaledger/goshimmer/packages/protocol/engine/clock"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/consensus"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/consensus/blockgadget"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/eviction"
@@ -49,7 +50,7 @@ type Engine struct {
 	Tangle              *tangle.Tangle
 	Consensus           *consensus.Consensus
 	TSCManager          *tsc.Manager
-	Clock               Clock
+	Clock               clock.Clock
 	SlotTimeProvider    *slot.TimeProvider
 
 	Workers *workerpool.Group
@@ -74,7 +75,7 @@ type Engine struct {
 func New(
 	workers *workerpool.Group,
 	storageInstance *storage.Storage,
-	clockProvider module.Provider[*Engine, Clock],
+	clockProvider module.Provider[*Engine, clock.Clock],
 	sybilProtectionProvider module.Provider[*Engine, sybilprotection.SybilProtection],
 	throughputQuotaProvider module.Provider[*Engine, throughputquota.ThroughputQuota],
 	slotTimeProvider *slot.TimeProvider,
@@ -181,7 +182,7 @@ func (e *Engine) IsBootstrapped() (isBootstrapped bool) {
 		return true
 	}
 
-	if isBootstrapped = time.Since(e.Clock.Acceptance().RelativeTime()) < e.optsBootstrappedThreshold && e.NotarizationManager.IsFullyCommitted(); isBootstrapped {
+	if isBootstrapped = time.Since(e.Clock.Accepted().RelativeTime()) < e.optsBootstrappedThreshold && e.NotarizationManager.IsFullyCommitted(); isBootstrapped {
 		e.isBootstrapped = true
 	}
 
@@ -189,7 +190,7 @@ func (e *Engine) IsBootstrapped() (isBootstrapped bool) {
 }
 
 func (e *Engine) IsSynced() (isBootstrapped bool) {
-	return e.IsBootstrapped() && time.Since(e.Clock.Acceptance().Time()) < e.optsBootstrappedThreshold
+	return e.IsBootstrapped() && time.Since(e.Clock.Accepted().Time()) < e.optsBootstrappedThreshold
 }
 
 func (e *Engine) Initialize(snapshot string) (err error) {
@@ -492,43 +493,5 @@ func WithRequesterOptions(opts ...options.Option[eventticker.EventTicker[models.
 		e.optsBlockRequester = append(e.optsBlockRequester, opts...)
 	}
 }
-
-// endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-// region Events ///////////////////////////////////////////////////////////////////////////////////////////////////////
-
-type Events struct {
-	Error          *event.Event1[error]
-	BlockProcessed *event.Event1[models.BlockID]
-
-	EvictionState       *eviction.Events
-	Filter              *filter.Events
-	Ledger              *ledger.Events
-	Tangle              *tangle.Events
-	Consensus           *consensus.Events
-	Clock               *ClockEvents
-	NotarizationManager *notarization.Events
-	SlotMutations       *notarization.SlotMutationsEvents
-	BlockRequester      *eventticker.Events[models.BlockID]
-
-	event.Group[Events, *Events]
-}
-
-// NewEvents contains the constructor of the Events object (it is generated by a generic factory).
-var NewEvents = event.CreateGroupConstructor(func() (newEvents *Events) {
-	return &Events{
-		Error:               event.New1[error](),
-		BlockProcessed:      event.New1[models.BlockID](),
-		EvictionState:       eviction.NewEvents(),
-		Filter:              filter.NewEvents(),
-		Ledger:              ledger.NewEvents(),
-		Tangle:              tangle.NewEvents(),
-		Consensus:           consensus.NewEvents(),
-		Clock:               NewClockEvents(),
-		NotarizationManager: notarization.NewEvents(),
-		SlotMutations:       notarization.NewSlotMutationsEvents(),
-		BlockRequester:      eventticker.NewEvents[models.BlockID](),
-	}
-})
 
 // endregion ///////////////////////////////////////////////////////////////////////////////////////////////////////////
