+++ packages/protocol/engine/testframework.go
@@ -7,6 +7,7 @@ import (
 
 	"github.com/iotaledger/goshimmer/packages/core/database"
 	"github.com/iotaledger/goshimmer/packages/core/module"
+	"github.com/iotaledger/goshimmer/packages/protocol/engine/clock"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/consensus/blockgadget"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/sybilprotection"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/tangle"
@@ -37,7 +38,7 @@ type TestFramework struct {
 	Acceptance    *blockgadget.TestFramework
 }
 
-func NewTestEngine(t *testing.T, workers *workerpool.Group, storage *storage.Storage, clock module.Provider[*Engine, Clock], sybilProtection module.Provider[*Engine, sybilprotection.SybilProtection], throughputQuota module.Provider[*Engine, throughputquota.ThroughputQuota], slotTimeProvider *slot.TimeProvider, opts ...options.Option[Engine]) *Engine {
+func NewTestEngine(t *testing.T, workers *workerpool.Group, storage *storage.Storage, clock module.Provider[*Engine, clock.Clock], sybilProtection module.Provider[*Engine, sybilprotection.SybilProtection], throughputQuota module.Provider[*Engine, throughputquota.ThroughputQuota], slotTimeProvider *slot.TimeProvider, opts ...options.Option[Engine]) *Engine {
 	e := New(workers.CreateGroup("Engine"), storage, clock, sybilProtection, throughputQuota, slotTimeProvider, opts...)
 	t.Cleanup(e.Shutdown)
 	return e
@@ -60,7 +61,7 @@ func NewTestFramework(test *testing.T, workers *workerpool.Group, engine *Engine
 	return t
 }
 
-func NewDefaultTestFramework(t *testing.T, workers *workerpool.Group, clock module.Provider[*Engine, Clock], sybilProtection module.Provider[*Engine, sybilprotection.SybilProtection], throughputQuota module.Provider[*Engine, throughputquota.ThroughputQuota], slotTimeProvider *slot.TimeProvider, optsEngine ...options.Option[Engine]) *TestFramework {
+func NewDefaultTestFramework(t *testing.T, workers *workerpool.Group, clock module.Provider[*Engine, clock.Clock], sybilProtection module.Provider[*Engine, sybilprotection.SybilProtection], throughputQuota module.Provider[*Engine, throughputquota.ThroughputQuota], slotTimeProvider *slot.TimeProvider, optsEngine ...options.Option[Engine]) *TestFramework {
 	engine := NewTestEngine(t, workers.CreateGroup("Engine"), blockdag.NewTestStorage(t, workers, database.WithDBProvider(database.NewDB)), clock, sybilProtection, throughputQuota, slotTimeProvider, optsEngine...)
 	t.Cleanup(engine.Shutdown)
