+++ packages/protocol/enginemanager/enginemanager.go
@@ -11,6 +11,7 @@ import (
 	"github.com/iotaledger/goshimmer/packages/core/database"
 	"github.com/iotaledger/goshimmer/packages/core/module"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine"
+	"github.com/iotaledger/goshimmer/packages/protocol/engine/clock"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/sybilprotection"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/throughputquota"
 	"github.com/iotaledger/goshimmer/packages/storage"
@@ -37,7 +38,7 @@ type EngineManager struct {
 	workers        *workerpool.Group
 
 	engineOptions           []options.Option[engine.Engine]
-	clockProvider           module.Provider[*engine.Engine, engine.Clock]
+	clockProvider           module.Provider[*engine.Engine, clock.Clock]
 	sybilProtectionProvider module.Provider[*engine.Engine, sybilprotection.SybilProtection]
 	throughputQuotaProvider module.Provider[*engine.Engine, throughputquota.ThroughputQuota]
 	slotTimeProvider        *slot.TimeProvider
@@ -51,7 +52,7 @@ func New(
 	dbVersion database.Version,
 	storageOptions []options.Option[database.Manager],
 	engineOptions []options.Option[engine.Engine],
-	clockProvider module.Provider[*engine.Engine, engine.Clock],
+	clockProvider module.Provider[*engine.Engine, clock.Clock],
 	sybilProtectionProvider module.Provider[*engine.Engine, sybilprotection.SybilProtection],
 	throughputQuotaProvider module.Provider[*engine.Engine, throughputquota.ThroughputQuota],
 	slotTimeProvider *slot.TimeProvider) *EngineManager {
