+++ packages/protocol/tipmanager/tipmanager.go
@@ -280,8 +280,8 @@ func (t *TipManager) isValidTip(tip *scheduler.Block) (err error) {
 	if !t.isPastConeTimestampCorrect(tip.Block) {
 		return errors.Errorf("cannot select tip due to TSC condition tip issuing time (%s), time (%s), min supported time (%s), block id (%s), tip pool size (%d), scheduled: (%t), orphaned: (%t), accepted: (%t)",
 			tip.IssuingTime(),
-			t.engine.Clock.Acceptance().Time(),
-			t.engine.Clock.Acceptance().Time().Add(-t.optsTimeSinceConfirmationThreshold),
+			t.engine.Clock.Accepted().Time(),
+			t.engine.Clock.Accepted().Time().Add(-t.optsTimeSinceConfirmationThreshold),
 			tip.ID().Base58(),
 			t.tips.Size(),
 			tip.IsScheduled(),
@@ -309,7 +309,7 @@ func (t *TipManager) IsPastConeTimestampCorrect(block *virtualvoting.Block) (tim
 //
 //	If there's any unaccepted block >TSC threshold, then the oldest accepted block will be >TSC threshold, too.
 func (t *TipManager) isPastConeTimestampCorrect(block *virtualvoting.Block) (timestampValid bool) {
-	minSupportedTimestamp := t.engine.Clock.Acceptance().Time().Add(-t.optsTimeSinceConfirmationThreshold)
+	minSupportedTimestamp := t.engine.Clock.Accepted().Time().Add(-t.optsTimeSinceConfirmationThreshold)
 
 	if !t.engine.IsBootstrapped() {
 		// If the node is not bootstrapped we do not have a valid timestamp to compare against.
