+++ plugins/remotemetrics/sync.go
@@ -23,10 +23,10 @@ func checkSynced() {
 			NodeID:         myID,
 			MetricsLevel:   Parameters.MetricsLevel,
 			Time:           time.Now(),
-			ATT:            deps.Protocol.Engine().Clock.Acceptance().Time(),
-			RATT:           deps.Protocol.Engine().Clock.Acceptance().RelativeTime(),
-			CTT:            deps.Protocol.Engine().Clock.Confirmation().Time(),
-			RCTT:           deps.Protocol.Engine().Clock.Confirmation().RelativeTime(),
+			ATT:            deps.Protocol.Engine().Clock.Accepted().Time(),
+			RATT:           deps.Protocol.Engine().Clock.Accepted().RelativeTime(),
+			CTT:            deps.Protocol.Engine().Clock.Confirmed().Time(),
+			RCTT:           deps.Protocol.Engine().Clock.Confirmed().RelativeTime(),
 			CurrentStatus:  tts,
 			PreviousStatus: oldTangleTimeSynced,
 		}
