+++ tools/genesis-snapshot/main.go
@@ -12,7 +12,7 @@ import (
 	"github.com/iotaledger/goshimmer/packages/core/snapshotcreator"
 	"github.com/iotaledger/goshimmer/packages/protocol"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine"
-	"github.com/iotaledger/goshimmer/packages/protocol/engine/clock"
+	"github.com/iotaledger/goshimmer/packages/protocol/engine/clock/blocktime"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/notarization"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/sybilprotection/dpos"
 	"github.com/iotaledger/goshimmer/packages/protocol/engine/throughputquota/mana1"
@@ -87,7 +87,7 @@ func diagnosticPrintSnapshotFromFile(filePath string, provider *slot.TimeProvide
 	s := createTempStorage()
 	defer s.Shutdown()
 
-	e := engine.New(workerpool.NewGroup("Diagnostics"), s, clock.Provide(), dpos.NewProvider(), mana1.NewProvider(), provider)
+	e := engine.New(workerpool.NewGroup("Diagnostics"), s, blocktime.NewProvider(), dpos.NewProvider(), mana1.NewProvider(), provider)
 	defer e.Shutdown()
 
 	if err := e.Initialize(filePath); err != nil {
